check if photos.json exists before attempting to delete it
git-svn-id: file:///root/var/svn/trunk/uploadr@428 2ccee060-bf41-0410-b5a3-d28fdbd3ea41

+++ MacUploadr.app/Contents/Resources/chrome/content/uploadr/file.js
@@ -78,7 +78,9 @@ var file = {
             var file = Cc['@mozilla.org/file/directory_service;1']
 				.getService(Ci.nsIProperties).get('ProfD', Ci.nsIFile);
 			file.append(name);
-			file.remove(false);
+			if (file.exists()) {
+			    file.remove(false);
+			}
         } catch (err) {
             Components.utils.reportError(err);
         }
